// frontend/src/client/types.gen.ts
export type ApartmentPublic = {
    address: string;
    city: string;
    area_sqm?: (number | null);
    rooms?: (number | null);
    floor?: (number | null);
    building_year?: (number | null);
    current_price: number;
    description?: (string | null);
    id: string;
    owner_id: string;
    created_at: string;
    updated_at: string;
};